Module governance::contracts

Functions

calc_total_voting_power_at

Calc total voting power at a specific time

calc_voting_power

Calc voting power for sender at a Proposal’s start_time

cast_vote

Cast a vote on an active proposal.

check_messages

Checks that proposal messages are correct. Returns ContractError on failure, otherwise returns a [Response] with the specified attributes. The last message will always fail to prevent committing into blockchain.

end_proposal

End a proposal and determine the result.

execute

execute_proposal

Execute Proposal Msgs

instantiate

query

query_proposal_voters

Return a list of voters for a given proposal

query_proposal_votes

Return the voting power per option for a given proposal

query_proposals

Return a list of Proposals

remove_completed_proposal

Remove completed Proposals

submit_proposal

Submit a proposal to the governance contract. Total stake must surpass the minimum. Only the vesting contract can submit expedited proposals.

update_config

Update the contract configuration