governance

Module governance

Structs

AutoExecuteToggled

Emitted when auto-execute is toggled for a proposal.

AutoExecuted

Emitted when a proposal is automatically executed.

CommentAdded

DelegationInfo

Delegation info for a governance signer.

DiscussionComment

A single comment or discussion entry on a governance proposal.

EmergencyOverrideUsed

Governance

GovernanceAnalytics

GovernanceProposal

GovernanceRef

ProposalCreated

ProposalExecuted

ProposalFromTemplate

Emitted when a proposal is created from a template.

ProposalRejected

PublicKeyRegistered

Emitted when a signer registers (or overwrites) their ECDSA public key.

QuadraticVote

Quadratic voting configuration for a proposal.

QuadraticVoteCast

SignerAdded

SignerRemoved

TemplateCreated

Emitted when a proposal template is created.

ThresholdUpdated

VoteCast

Enums

Error

GovernanceAction

ProposalStatus